Split @wordpress/url into modules.
