+++ packages/url/src/is-valid-authority.js
@@ -1,0 +1,19 @@
+/**
+ * Checks for invalid characters within the provided authority.
+ *
+ * @param {string} authority A string containing the URL authority.
+ *
+ * @example
+ * ```js
+ * const isValid = isValidAuthority( 'wordpress.org' ); // true
+ * const isNotValid = isValidAuthority( 'wordpress#org' ); // false
+ * ```
+ *
+ * @return {boolean} True if the argument contains a valid authority.
+ */
+export function isValidAuthority( authority ) {
+	if ( ! authority ) {
+		return false;
+	}
+	return /^[^\s#?]+$/.test( authority );
+}
