+++ packages/url/src/prepend-http.js
@@ -1,0 +1,31 @@
+/**
+ * Internal dependencies
+ */
+import { isEmail } from './is-email';
+
+const USABLE_HREF_REGEXP = /^(?:[a-z]+:|#|\?|\.|\/)/i;
+
+/**
+ * Prepends "http://" to a url, if it looks like something that is meant to be a TLD.
+ *
+ * @param {string} url The URL to test.
+ *
+ * @example
+ * ```js
+ * const actualURL = prependHTTP( 'wordpress.org' ); // http://wordpress.org
+ * ```
+ *
+ * @return {string} The updated URL.
+ */
+export function prependHTTP( url ) {
+	if ( ! url ) {
+		return url;
+	}
+
+	url = url.trim();
+	if ( ! USABLE_HREF_REGEXP.test( url ) && ! isEmail( url ) ) {
+		return 'http://' + url;
+	}
+
+	return url;
+}
